pyproject.toml
[tool.poetry]
name = "tsp-microcontroller-application"
version = "2023.03.16.00"
description = "The application to run on the microcontroller of our smart home system"
authors = [
    "J or Jacob Babich <jacobbabichpublic+git@gmail.com>",
    "Evan Cardasis <todoaddtheiremails@example.com>",
    "Robert Hausz <todoaddtheiremails@example.com>",
    "Kevin Mexicotte <todoaddtheiremails@example.com>",
]
readme = "README.md"
packages = [{ include = "microcontroller_application" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
# apache-beam = "^2.45.0"
babichjacob-emitter = "^0.1.1"
babichjacob-emitter-tools = "^0.1.2"
babichjacob-bounded-channel = "^0.2.0"
babichjacob-oneshot-channel = "^0.2.0"
babichjacob-option-and-result = "^0.1.2"
babichjacob-store = "^0.2.0"
# dlib = "^19.24.0"
# face-recognition = "^1.3.0"
numpy = "1.22.4"
# object-detection = { path = "./models/research", develop = true }
opencv-python = "^4.7.0.72"
pillow = "^9.4.0"
pandas = "^1.5.3"
# tensorflow = "^2.11.0"
# tensorflow-datasets = "^4.8.3"
# tf-models-official = "^2.11.3"
websockets = "^10.4"
msgpack = "^1.0.5"

[tool.poetry.dev-dependencies]
types-Pillow = "^9.4.0.13"
black = "^22.12.0"
mypy = "^0.991"
pylint = "^2.15.10"

[tool.poetry.group.test.dependencies]
imagehash = "^4.3.1"
pytest = "^7"
pytest-asyncio = "^0.20.3"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.21.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
